Fix poor SID time handling
We did not handle the case where a SID is active only on certain
weekdays and the activation and deactivation are on the same day.
This caused mistaken selection of CIV2D.

Further the order of SID filter operations has been optimised
and a help message for the force-sid command is added.